combine related tests from separate files
